pyproject.toml
[tool.poetry]
name = "Skin-Cancer-Classification"
version = "0.1.0"
description = "mlops project"
authors = ["Kseniia Karnakova <karnakovakk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.2"
torchvision = "^0.17.2"
pytorch-lightning = "^2.4.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
mlflow = "^2.19.0"
dvc = "^3.58.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
